+++ packages/xchain-bsc/CHANGELOG.md
@@ -1,6 +1,13 @@
+# v.0.1.3 (2023-04-05)
+
+## Add
+
+- bump xchain-client deps
+
 # v.0.1.2 (2023-03-21)
 
 ## Update
+
 - Update explorer provider import
 
 # v.0.1.1 (2023-03-18)